requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.2.4
yfinance==0.2.55
plotly==6.0.1
python-dotenv==1.1.0
requests==2.32.3
